- Primeira implementação funcional de services com filtros dinâmicos baseados nos DTO's. - Implementação de novas interfaces para permitir maior generalização de filtros - Implementação de Helper na montagem de expressions para uso nos filtros dinâmicos. - Adicionado logger de arquivo - Alterado o tipo base das entidades para long - Removida CS do web.config - Implementado BaseRestController - Adicionadas extensões de IQueryables que facilitam a extensão de funcionalidades de conversão LINQ > Nhibernate

+++ EnadeProject.NHibernate/NHibernate/EntityMappings/FrameWork_Entities/EntidadeBase.cs
@@ -6,9 +6,9 @@ namespace EnadeProject.NHibernate.EntityMappings.FrameWork_Entities
 {
     /// <summary>
     ///     É preciso criar uma entidade base sem tipo para que não ser necessário replicar o tipo de entidade base para todas
-    ///     entidades
+    ///     entidades. Desta Maneira, alinhamos os tipos de dados dos identificadores burros (PK sequencial).
     /// </summary>
-    public abstract class EntidadeBase : Entity<int>
+    public abstract class EntidadeBase : Entity<long>
     {
     }
 }
